Adds bootstrap
I added the bootstrap library to make it look half-way decent.  I also
applied it to all of the views

The only other change that I made was to scope each of the actions in
the polls controller to the current user so that people won't be able to
modify each others polls.