Adds options to poll controller
This change introduces the Option model into the Poll controller.  I've
made sure to add Options to every action that interacts with a Poll.
Mainly this consists of being sure to `include` the options when looking
up a poll to reduce the number of queries.  This also means that I've
added `new` Options when creating Polls.

I'm not 100% sure this will work when editing a poll.  I need to add
some tests around it to find out.