Fix testing setup
Until now, I hadn't been testing any thing that I was doing.  I was just
making changes and seeing if the worked.  That's not really sustainable
for long term development so I got everything setup according to
[Thoughtbot](https://robots.thoughtbot.com/how-we-test-rails-applications).

Just to save a trip to that link, the libraries I'm using are:

- rspec: Generally for all tests
- capybara and capybara-webkit: for integration tests with javascript
- factory_girl: Using factories instead of fixtures

The main exception to the libraries that thoughtbot lists is that I
didn't include webmock since I don't think I'll have any external calls.
If I need it, I can always add it in later.