Fixes response index controller test
This changes the test to actually test something instead of just
skipping it.  Currently it just checks that it returns a 200.  I think
that it should probably test more but I'm having trouble figuring out
how in Rails 5.  Most of the documentation on controller testing with
RSpec is for Rails 4 which is apparently wildly different.  I guess some
test is better than no test though, right future Chris?